pyproject.toml
[project]
name = "rwa-tokenization"
version = "0.1.0"
description = "RWA tokenization framework with Postgres and Kafka"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary==2.9.11",
    "kafka-python-ng==2.2.3",
]

[project.optional-dependencies]
test = [
    "pytest==8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
